# src/main/resources/contract/openapi.yaml
openapi: 3.0.3
info:
  title: JarFactory API
  version: 1.0.0
  description: API for ordering jars from the JarFactory

servers:
  - url: http://localhost:8080/api/v1
    description: Jam server

tags:
  - name: Orders
    description: Endpoints related to ordering jars.

paths:
  /jars/order:
    post:
      tags:
        - Orders
      summary: Create a new jar order
      operationId: createJarOrder
      requestBody:
        description: Order details for jars
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJarOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Business logic error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessErrorResponse'
        '501':
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'

    put:
      tags:
        - Orders
      summary: Update the delivery date and contents of an existing jar order
      operationId: updateJarOrder
      requestBody:
        description: Updated order details for jars
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJarOrderRequest'
      responses:
        '204':
          description: Order updated successfully, no content returned
        '501':
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'

components:
  schemas:
    CreateJarOrderRequest:
      type: object
      properties:
        deliveryDate:
          type: string
          format: date
          example: "2024-08-01"
        smallJars:
          type: integer
          example: 10
        mediumJars:
          type: integer
          example: 5
        largeJars:
          type: integer
          example: 2
      required:
        - deliveryDate
        - smallJars
        - mediumJars
        - largeJars

    UpdateJarOrderRequest:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        deliveryDate:
          type: string
          format: date
          example: "2024-08-10"
        smallJars:
          type: integer
          example: 15
        mediumJars:
          type: integer
          example: 8
        largeJars:
          type: integer
          example: 3
      required:
        - orderId
        - deliveryDate
        - smallJars
        - mediumJars
        - largeJars

    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"

    BusinessErrorResponse:
      type: object
      properties:
        rejectReason:
          type: string
          example: "zamówienie przekraczajace zdolności produkcyjne"

    NotImplementedError:
      type: object
      properties:
        error:
          type: string
          example: "Not Implemented"